requirements.txt
pytz
django~=3.2.8
django-filter~=21.1

django-allauth~=0.45.0
python-dateutil~=2.8.2

markdown
django-filter
django-resized
python-decouple
django-ckeditor
django-crispy-forms~=1.13.0
pillow
six~=1.16.0

# python -m pip install --upgrade Pillow >>> install pillow with this
